move TIMEOUTS out of fixtures and into support/constants
